+++ src/readfiles/codeblocks.rs
@@ -2,15 +2,18 @@ use crate::syntax_hl::ClassedHTMLGenerator;
 use crate::syntax_hl::LinesWithEndings;
 use anyhow::{bail, Result};
 use pulldown_cmark::escape::escape_html;
+use qr_code::QrCode;
+use std::fmt::Write;
 
 pub trait BlockHandler {
     fn push(&mut self, content: &str) -> Result<()>;
-    fn end(self);
+    fn end(self) -> Result<()>;
 }
 
 pub enum DynBlock<'a> {
     Leaflet(LeafletHandler<'a>),
     Code(CodeBlock<'a>),
+    Qr(QrHandler<'a>),
 }
 impl<'a> DynBlock<'a> {
     pub fn for_kind(
@@ -21,6 +24,7 @@ impl<'a> DynBlock<'a> {
             Some("leaflet") => {
                 Ok(DynBlock::Leaflet(LeafletHandler::open(out)))
             }
+            Some("qr") => Ok(DynBlock::Qr(QrHandler::open(out))),
             Some(bang) => {
                 bail!("Magic for !{:?} not implemented", bang);
             }
@@ -34,12 +38,14 @@ impl<'a> BlockHandler for DynBlock<'a> {
         match self {
             DynBlock::Leaflet(x) => x.push(content),
             DynBlock::Code(x) => x.push(content),
+            DynBlock::Qr(x) => x.push(content),
         }
     }
-    fn end(self) {
+    fn end(self) -> Result<()> {
         match self {
             DynBlock::Leaflet(x) => x.end(),
             DynBlock::Code(x) => x.end(),
+            DynBlock::Qr(x) => x.end(),
         }
     }
 }
@@ -70,8 +76,63 @@ impl<'a> BlockHandler for LeafletHandler<'a> {
         Ok(())
     }
 
-    fn end(self) {
+    fn end(self) -> Result<()> {
         self.out.push_str("}\n</script>\n");
+        Ok(())
+    }
+}
+
+pub struct QrHandler<'a> {
+    out: &'a mut String,
+    data: String,
+}
+
+impl<'a> QrHandler<'a> {
+    fn open(out: &'a mut String) -> Self {
+        QrHandler {
+            out,
+            data: String::new(),
+        }
+    }
+}
+impl<'a> BlockHandler for QrHandler<'a> {
+    fn push(&mut self, content: &str) -> Result<()> {
+        self.data.push_str(content);
+        Ok(())
+    }
+
+    fn end(self) -> Result<()> {
+        let qr = QrCode::new(self.data)?;
+        let width = qr.width();
+        let mut imgdata = Vec::new();
+        let mut img = png::Encoder::new(&mut imgdata, width as _, width as _);
+        img.set_color(png::ColorType::Grayscale);
+        img.set_depth(png::BitDepth::One);
+        img.set_compression(png::Compression::Best);
+        let mut writer = img.write_header()?;
+        writer.write_image_data(
+            &qr.to_vec()
+                .chunks(width)
+                .flat_map(|line| {
+                    line.chunks(8).map(|byte| {
+                        byte.iter()
+                            .fold(0, |acc, elem| 2 * acc + u8::from(!elem))
+                            * (1 << (8 - byte.len()))
+                    })
+                })
+                .collect::<Vec<_>>(),
+        )?;
+        writer.finish()?;
+
+        let url =
+            format!("data:image/png;base64,{}", base64::encode(imgdata));
+        writeln!(
+            self.out,
+            "<figure class='qr-code sidebar'>\
+             <img alt='' src='{url}' width='{width}' height='{width}'/>\
+             </figure>",
+        )?;
+        Ok(())
     }
 }
 
@@ -108,10 +169,11 @@ impl<'a> BlockHandler for CodeBlock<'a> {
         }
         Ok(())
     }
-    fn end(self) {
+    fn end(self) -> Result<()> {
         if let Some(gen) = self.gen {
             self.out.push_str(&gen.finalize())
         }
         self.out.push_str("</pre>\n");
+        Ok(())
     }
 }